Add AuthController and UserController in API layer

+++ CodeBook.Business.App/Services/UserService.cs
@@ -20,21 +20,21 @@ namespace CodeBook.Business.App.Services
             _followRepository = followRepository;
             this.mapper = mapper;
         }
-        public void DeleteAccount(int userId) 
+        public bool DeleteAccount(int userId) 
         {
             User user = _userRepository.GetProfileById(userId);
             
             _userRepository.Remove(user);
-            _userRepository.SaveChanges();
+           return _userRepository.SaveChanges();
 
         }
-        public UserProfileResponse GetProfile(int userId) 
+        public UserProfileResponse GetProfile(int userId)
         {
             User user = _userRepository.GetProfileById(userId);
-            return mapper.Map<UserProfileResponse>(user);
-            //return user;        
+            if (user == null) return null;
+            return mapper.Map<UserProfileResponse>(user);     
         }
-        public void UpdateProfile(int userId,UpdateProfileDto data) 
+        public bool UpdateProfile(int userId,UpdateProfileDto data) 
         {
             User user = _userRepository.GetProfileById(userId);
 
@@ -42,7 +42,7 @@ namespace CodeBook.Business.App.Services
             user.UserName = data.UserName;
             user.AvatarUrl = data.AvatarUrl;
             _userRepository.Update(user);
-            _userRepository.SaveChanges();
+            return _userRepository.SaveChanges();
 
         }
         public bool VerifyPassword(string password,int userId)
@@ -51,23 +51,24 @@ namespace CodeBook.Business.App.Services
 
             return BCrypt.Net.BCrypt.Verify(password, user.PasswordHash);
         }
-        public void Follow(int followerId, int followeeId)
+        public bool Follow(int followerId, int followeeId)
         {
+            Follow followedalready = _followRepository.GetFollow(followerId, followeeId);
+            if (followedalready != null) return false;
             Follow follow = new Follow();
             follow.FollowerUserId = followerId;
             follow.FolloweeUserId = followeeId;
             _followRepository.AddFollow(follow);
-            _followRepository.SaveChanges();
-
+           return _followRepository.SaveChanges();
 
         }
 
-        public void Unfollow(int followerId, int followeeId)
+        public bool Unfollow(int followerId, int followeeId)
         {
             Follow follow = _followRepository.GetFollow(followerId, followeeId);
-
+            if(follow == null) return false;
             _followRepository.RemoveFollow(follow);
-            _followRepository.SaveChanges();
+            return _followRepository.SaveChanges();
 
         }
     }
